// src/domain/profile/profile.enums.ts
export enum Gender {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
  OTHER = 'OTHER',
  PREFER_NOT_TO_SAY = 'PREFER_NOT_TO_SAY',
}

export enum FitnessGoal {
  LOSE_FAT = 'LOSE_FAT',
  BUILD_MUSCLE = 'BUILD_MUSCLE',
  STRENGTH = 'STRENGTH',
  ENDURANCE = 'ENDURANCE',
  GENERAL_FITNESS = 'GENERAL_FITNESS',
  RECOMPOSITION = 'RECOMPOSITION',
}

export enum ActivityLevel {
  SEDENTARY = 'SEDENTARY',
  LIGHT = 'LIGHT',
  MODERATE = 'MODERATE',
  ACTIVE = 'ACTIVE',
  VERY_ACTIVE = 'VERY_ACTIVE',
}

export enum WeightUnit {
  KG = 'KG',
  LB = 'LB',
}

export enum HeightUnit {
  CM = 'CM',
  IN = 'IN',
}